+++ internal/froxysvr/httpforward/forward.go
@@ -1,4 +1,4 @@
-package forward
+package httpforward
 
 import (
 	"net/http"
