+++ internal/froxysvr/httpforward/mitm.go
@@ -1,4 +1,4 @@
-package forward
+package httpforward
 
 import (
 	"bufio"
