+++ internal/froxysvr/httpforward/plain.go
@@ -1,4 +1,4 @@
-package forward
+package httpforward
 
 import (
 	"fmt"
