+++ internal/froxysvr/httpforward/standard.go
@@ -1,4 +1,4 @@
-package forward
+package httpforward
 
 import (
 	"log"
