+++ internal/froxysvr/httpforward/tunnel.go
@@ -1,4 +1,4 @@
-package forward
+package httpforward
 
 import (
 	"io"
