+++ internal/froxysvr/httpreverse/roundrobin.go
@@ -1,4 +1,4 @@
-package reverse
+package httpreverse
 
 import (
 	"context"
